Upgrade dbt core APIs to 1.9.x - first pass.
This commit is the first pass at upgrading to the latest version of the dbt
core APIs for the adapter. Once this is complete and compiles cleanly, we'll
do subsequent passes to knock out runtime and semantic errors that remain.

As of this commit, `poetry run pytest` at least compiles and executes with
most tests passing. More work is needed, however. See below.

This includes the following major changes:

* A number of classes and types have moved package locations. I've done my best
  to make sure the replacements selected are semantically equivalent, but I'm
  not positive that's the case.  For this pass, I'm targeting a clean compile
  (with type checking enabled) to catch the majority of obvious mistakes but it
  deserves a second pass.
* Some of the error/exception names have changed, and some have no obvious
  replacement. I haven't done a deep dive to understand if the error types
  effect different error handling logic or if they're just used for semantic
  differentiation. This also deserves a second pass.
* The custom implementation of BaseRelation is temporarily commented out. See
  the larger comment below on this. I've also left this comment in a comment
  in case someone only looks at the code.
* DecodableAdapter had an implementation of AdapterProtocol that doesn't seem
  to be required anymore. I've commented it out and left it in place because
  I can't find any documentation on this change from dbt. I suspect this will
  cause some kind of issue once everything is running.
* Some methods of BaseAdapter are now allowed to simply `pass` rather than return
  NotImplementedException (which no longer exists). expand_column_types() is
  one place that got this treatment.
* get_thread_connection() has moved from BaseAdapter to BaseAdapter#connections.
* DecodableRelation is currently unused due to the dataclass issue described above.

On the DecodableRelation / frozen dataclass issue:

For some reason the base class is calling cls.from_dict(kwargs) in create() on
our implementation of Relation. This fails with an AttributeError because
Relation is marked as a frozen dataclass (see @dataclass def) which is required
because BaseRelation is also frozen. This means the base class is trying to
mutate something it knows must be immutable which makes no sense. There is a
deeper issue here that needs to be worked out.  That said, it's not clear we
require a custom implementation of BaseRelation.  I've temporarily updated many
of the methods that were referencing Relation to use BaseRelation instead since
they didn't appear to touch any specific members of Relation. It's all very
strange.